Portfolio Changes

+++ src/app/components/photo-uploader/photo-uploader.component.ts
@@ -1,4 +1,4 @@
-import { NgIf } from '@angular/common';
+import { NgClass, NgIf } from '@angular/common';
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { ImageCroppedEvent, ImageCropperModule } from 'ngx-image-cropper';
 import { ProfilePhoto } from 'src/app/common/profile-photo';
@@ -6,15 +6,16 @@ import { ProfilePhotoService } from 'src/app/services/profile-photo.service';
 import { constants } from 'src/environments/constants';
 
 @Component({
-    selector: 'app-photo-uploader',
-    templateUrl: './photo-uploader.component.html',
-    styleUrls: ['./photo-uploader.component.css'],
-    standalone: true,
-    imports: [NgIf, ImageCropperModule]
+  selector: 'app-photo-uploader',
+  templateUrl: './photo-uploader.component.html',
+  styleUrls: ['./photo-uploader.component.css'],
+  standalone: true,
+  imports: [NgIf, ImageCropperModule, NgClass]
 })
-export class PhotoUploaderComponent implements OnInit{
+export class PhotoUploaderComponent implements OnInit {
 
   @Input() customerId: number = 0;
+  @Input() listingId: number = 0;
   @Output() reloadEvent = new EventEmitter<boolean>();
   base64Data: any = '';
 
@@ -22,7 +23,7 @@ export class PhotoUploaderComponent implements OnInit{
   imgChangeEvt: any = '';
   cropImgPreview: any = '';
 
-  selectedFile!:Blob;
+  selectedFile!: Blob;
   profilePhoto!: ProfilePhoto;
 
   constructor(
@@ -48,53 +49,69 @@ export class PhotoUploaderComponent implements OnInit{
 
   //Gets called when the user clicks on retieve image button to get the image from back end
   getImage() {
-    const subscription = this.profilePhotoService.getImageByCustomerId(this.customerId, true).subscribe(
-      (image) => {
-        if(image&&image.state!=constants.ERROR_STATE){
-        this.base64Data = image.picByte;
-        this.profilePhoto = image;
-        this.profilePhoto.picByte = 'data:image/jpeg;base64,' + image.picByte;
-      }
+    //check if photo uploader or portfolio
+    if (this.customerId > 0) {
+      const subscription = this.profilePhotoService.getImageByCustomerId(this.customerId, true).subscribe(
+        (image) => {
+          if (image && image.state != constants.ERROR_STATE) {
+            this.base64Data = image.picByte;
+            this.profilePhoto = image;
+            this.profilePhoto.picByte = 'data:image/jpeg;base64,' + image.picByte;
+          }
 
-        subscription.unsubscribe();
-      }
-    );
+          subscription.unsubscribe();
+        }
+      );
+    } else if (this.listingId > 0) {
+
+    }
+
   }
 
   //Gets called when the user clicks on submit to upload the image
   uploadPhoto() {
 
     let uploadImageData = new FormData();
-    uploadImageData.append('imageFile', this.selectedFile, ""+this.customerId);
+    uploadImageData.append('imageFile', this.selectedFile, "" + this.customerId);
 
-    const subscription = this.profilePhotoService.uploadImage(uploadImageData).subscribe(
-      (response) => {
-        if(response.state!=constants.ERROR_STATE) {
-          this.reloadCurrentPage();
+    //check if photo uploader or portfolio
+    if (this.customerId > 0) {
+      const subscription = this.profilePhotoService.uploadImage(uploadImageData).subscribe(
+        (response) => {
+          if (response.state != constants.ERROR_STATE) {
+            this.reloadCurrentPage();
+          }
+          subscription.unsubscribe();
         }
-        subscription.unsubscribe();
-      }
-    );
+      );
+    } else if (this.listingId > 0) {
+
+    }
 
   }
 
   removePhoto() {
-    const subscription = this.profilePhotoService.removeImage().subscribe(
-      (response) => {
-        if(response.state!=constants.ERROR_STATE) {
-          this.reloadCurrentPage();
+    //check if photo uploader or portfolio
+    if (this.customerId > 0) {
+      const subscription = this.profilePhotoService.removeImage().subscribe(
+        (response) => {
+          if (response.state != constants.ERROR_STATE) {
+            this.reloadCurrentPage();
+          }
+          subscription.unsubscribe();
         }
-        subscription.unsubscribe();
-      }
-    );
+      );
+    } else if (this.listingId > 0) {
+
+    }
   }
 
   reloadCurrentPage() {
     window.location.reload();
-   }
+  }
 
-   reloadComponent() {
+  reloadComponent() {
     this.reloadEvent.emit(true);
-   }
+  }
 
 }
